requirements.txt
beautifulsoup4==4.11.1
python-telegram==0.15.0
requests==2.22.0
requests-unixsocket==0.2.0
google-images-download==2.8.0